% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_matrix.R
\name{calculate_gamma_hat}
\alias{calculate_gamma_hat}
\title{Calcualte the maximum stable gamma}
\usage{
calculate_gamma_hat(web, rho, delta)
}
\arguments{
\item{web}{binary network structure}

\item{rho}{mean field interspecific competition}

\item{delta}{trade-off}
}
\value{
stability condition
}
\description{
Calcualte the maximum stable gamma
}
