% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_gradient.R
\name{calculate_omega}
\alias{calculate_omega}
\title{calculate structural stability of an interaction network}
\usage{
calculate_omega(alpha)
}
\arguments{
\item{alpha}{parametrized network structure}
}
\value{
structural stability
}
\description{
calculate structural stability of an interaction network
}
