% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_motif_gradient.R
\name{check_motif_feasibility}
\alias{check_motif_feasibility}
\title{check if motif is feasible}
\usage{
check_motif_feasibility(alpha, r, label)
}
\value{
motif feasibility
}
\description{
check if motif is feasible
}
