% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_matrix.R
\name{construct_random_interaction_component}
\alias{construct_random_interaction_component}
\title{Construct a random matrix}
\usage{
construct_random_interaction_component(web, strength = 1, interaction_type)
}
\arguments{
\item{web}{binary network structure}

\item{strength}{average interaction strength}

\item{interaction_type}{mutualistic or antagonistic}
}
\value{
paramatrized interaction matrix
}
\description{
Construct a random matrix
}
