% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig3B.R
\name{generate_figure3B_data}
\alias{generate_figure3B_data}
\title{Reproduce the data used in Figure 3B in the main text}
\usage{
generate_figure3B_data()
}
\description{
Reproduce the data used in Figure 3B in the main text
}
