% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig3C.R
\name{generate_figure3C_data}
\alias{generate_figure3C_data}
\title{Reproduce the data used in Figure 3C in the main text}
\usage{
generate_figure3C_data()
}
\description{
Reproduce the data used in Figure 3C in the main text
}
