% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig5.R
\name{generate_figure_5_data}
\alias{generate_figure_5_data}
\title{Reproduce the data used in Figure 5 in the main text}
\usage{
generate_figure_5_data()
}
\description{
Reproduce the data used in Figure 5 in the main text
}
