% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_gradient.R
\name{get_motif_gradient}
\alias{get_motif_gradient}
\title{calculate motif profiles with a list of intrinic
growth rates}
\usage{
get_motif_gradient(Inte, r_set, bipartite)
}
\arguments{
\item{Inte}{parametrized network structure}
}
\value{
a tibble
}
\description{
calculate motif profiles with a list of intrinic
growth rates
}
