% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_motif_gradient.R
\name{get_subnetwork_gradient}
\alias{get_subnetwork_gradient}
\title{calculate motif profiles with a list of intrinic
growth rates for subnetworks}
\usage{
get_subnetwork_gradient(Inte, r_set, bipartite)
}
\arguments{
\item{Inte}{parametrized network structure}

\item{r_set}{intrinsic growth rates}

\item{bipartite}{bipartite network structure}
}
\value{
a tibble
}
\description{
calculate motif profiles with a list of intrinic
growth rates for subnetworks
}
