% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_motif_gradient.R
\name{identify_motif}
\alias{identify_motif}
\title{identify the type of a motif}
\usage{
identify_motif(row_label, col_label, bipartite)
}
\value{
motif type
}
\description{
identify the type of a motif
}
